fix: rename OpenAIHelper, remove OpenAIProvider, fix LDMessage[] input handling
- Rename openaiHelper.ts → OpenAIHelper.ts to match PascalCase convention
- Delete OpenAIProvider.ts (superseded by OpenAIRunnerFactory + OpenAIHelper)
- OpenAIModelRunner.run() now accepts string | LDMessage[]; passes LDMessage[]
  directly to the API so judge evaluation works correctly
- OpenAIRunnerFactory constructor takes only logger; instantiates OpenAI client
  internally rather than accepting an injected client
