+++ packages/ai-providers/server-ai-openai/src/OpenAIHelper.ts
