+++ packages/ai-providers/server-ai-openai/src/OpenAIModelRunner.ts
@@ -8,7 +8,7 @@ import type {
   RunnerResult,
 } from '@launchdarkly/server-sdk-ai';
 
-import { convertMessagesToOpenAI, getAIMetricsFromResponse } from './openaiHelper';
+import { convertMessagesToOpenAI, getAIMetricsFromResponse } from './OpenAIHelper';
 
 /**
  * Runner implementation for OpenAI chat completions.
@@ -32,17 +32,20 @@ export class OpenAIModelRunner implements Runner {
   }
 
   /**
-   * Run the OpenAI model with the given prompt.
+   * Run the OpenAI model with the given prompt or message array.
    *
-   * @param input The user prompt to send to the model.
+   * When `input` is a string it is wrapped as a user turn and appended to any
+   * messages defined in the config. When `input` is already a `LDMessage[]`
+   * (e.g. when called from the Judge evaluation path) it is used as-is.
+   *
+   * @param input The user prompt string, or a pre-built message array.
    * @param outputType Optional JSON schema for structured output. When provided,
    *   the response is parsed and exposed via {@link RunnerResult.parsed}.
    */
-  async run(input: string, outputType?: Record<string, unknown>): Promise<RunnerResult> {
-    const messages: LDMessage[] = [
-      ...(this._config.messages ?? []),
-      { role: 'user', content: input },
-    ];
+  async run(input: string | LDMessage[], outputType?: Record<string, unknown>): Promise<RunnerResult> {
+    const messages: LDMessage[] = Array.isArray(input)
+      ? input
+      : [...(this._config.messages ?? []), { role: 'user', content: input }];
 
     if (outputType !== undefined) {
       return this._runStructured(messages, outputType);
