+++ packages/ai-providers/server-ai-openai/src/index.ts
@@ -1,4 +1,3 @@
-export { OpenAIProvider } from './OpenAIProvider';
 export { OpenAIModelRunner } from './OpenAIModelRunner';
 export { OpenAIAgentRunner, ToolRegistry } from './OpenAIAgentRunner';
 export { OpenAIRunnerFactory } from './OpenAIRunnerFactory';
@@ -6,5 +5,5 @@ export {
   convertMessagesToOpenAI,
   getAIMetricsFromResponse,
   getAIUsageFromResponse,
-} from './openaiHelper';
-export type { OpenAIChatMessage } from './openaiHelper';
+} from './OpenAIHelper';
+export type { OpenAIChatMessage } from './OpenAIHelper';
